refactor(interfaces): move general and environment interfaces to shared folder
Consolidate common interfaces into a shared directory to improve code organization and maintainability. This change updates all import paths to reflect the new location.
