+++ src/domain/auth/dtos/register.dto.ts
@@ -1,5 +1,5 @@
 import { ERRORS } from "@/config/strings/global.strings.json";
-import { TRawJson } from "../../interfaces/general.interfaces";
+import { TRawJson } from "../../shared/interfaces/general.interfaces";
 
 interface IRegisterDto {
   name: string;
