+++ src/domain/dtos/encrypted-body.dto.ts
@@ -3,7 +3,7 @@ import {
   IEncryption,
   ISignature,
 } from "../interfaces/encrypted-body.interfaces";
-import { TRawJson } from "../interfaces/general.interfaces";
+import { TRawJson } from "../shared/interfaces/general.interfaces";
 
 import { ERRORS } from "@/config/strings/global.strings.json";
