+++ src/domain/log/dtos/create-log.dto.ts
@@ -1,5 +1,5 @@
 import { ERRORS } from "@/config/strings/global.strings.json";
-import { TRawJson } from "@/domain/interfaces/general.interfaces";
+import { TRawJson } from "@/domain/shared/interfaces/general.interfaces";
 import { LogLevel, LOG_LEVELS } from "@/domain/log/interfaces/log.interfaces";
 
 interface ICreateLogDto {
