+++ src/domain/shared/interfaces/environtments.interfaces.ts
