+++ src/domain/shared/interfaces/general.interfaces.ts
