+++ src/infrastructure/datasources/auth.datasource.ts
@@ -12,7 +12,7 @@ import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { UpdateUserDto } from "@/domain/user/dtos/update-user.dto";
 import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
 import { SupabaseClient } from "@supabase/supabase-js";
-import { TRawJson } from "@/domain/interfaces/general.interfaces";
+import { TRawJson } from "@/domain/shared/interfaces/general.interfaces";
 
 export class AuthDatasource implements AuthRepository {
   constructor(private readonly client: typeof AuthClient) {
