+++ src/infrastructure/validators/request-reset-password-email.validator.ts
@@ -3,7 +3,7 @@ import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-p
 import globalStrings from "@/config/strings/global.strings.json";
 import { EMAIL_BASIC_REGEX, URL_REGEX } from "@/config/regex/validations.regex";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
-import { TRawJson } from "@/domain/interfaces/general.interfaces";
+import { TRawJson } from "@/domain/shared/interfaces/general.interfaces";
 import { processValidationError } from "./utils/processError.validator";
 
 const { VALIDATION } = globalStrings.ERRORS.AUTH.REQUEST_RESET_PASSWORD_EMAIL;
